# ecommerce/openapi.yaml
openapi: 3.0.3
info:
  title: 오늘 한 차 (OneulHanCha) API
  description: |
    # 오늘 한 차 - 티 박스 구독 서비스 API
    
    나의 컨디션, 원하는 기분, 선호하는 향에 맞춰 큐레이션된 7일 티 박스를 제공하는 서비스입니다.
    
    ## 주요 기능
    - 🫖 박스 상품 조회 및 조합 선택
    - 🛒 장바구니 관리
    - 📦 주문 및 결제 (포인트)
    - 🎟️ 선착순 쿠폰 발급
    - 💰 포인트 충전 및 관리
    
    ## 인증
    JWT Bearer 토큰을 사용합니다.
    
    ## Rate Limiting
    - 전체 API: 100 요청/분/사용자
    - 쿠폰 발급: 5 요청/분/사용자
    - 주문 생성: 10 요청/분/사용자
    
    ## 에러 처리
    모든 에러는 표준화된 형식으로 반환됩니다.

  version: 1.0.0
  contact:
    name: API Support
    email: support@oneulhancha.com
  license:
    name: Proprietary
servers:
  - url: https://api.oneulhancha.com/api/v1
    description: 운영 서버
  - url: https://dev-api.oneulhancha.com/api/v1
    description: 개발 서버
  - url: http://localhost:8080/api/v1
    description: 로컬 서버
...
security:
  - BearerAuth: []
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: 주문 내역 조회
      description: 사용자의 주문 내역을 페이징하여 조회합니다.
      operationId: getOrders
      parameters:
        - name: page
          in: query
          description: 페이지 번호 (0부터 시작)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          example: 0
        - name: size
          in: query
          description: 페이지 크기
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          example: 10
        - name: status
          in: query
          description: 주문 상태 필터
          required: false
          schema:
            type: string
            enum: [ALL, PENDING, PAID, PREPARING, SHIPPED, DELIVERED, CANCELLED]
            default: ALL
          example: ALL
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Orders
      summary: 주문 생성
      description: 장바구니의 상품을 주문하고 결제합니다.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              couponId: 1
              deliveryAddress:
                recipient: "김민지"
                phone: "010-1234-5678"
                zipCode: "06234"
                address: "서울시 강남구 테헤란로 123"
                addressDetail: "456호"
      responses:
        '201':
          description: 주문 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 충돌 (재고 부족, 포인트 부족)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                inventoryShortage:
                  summary: 재고 부족
                  value:
                    success: false
                    error:
                      code: "INVENTORY002"
                      message: "재고가 부족하여 주문할 수 없습니다"
                    timestamp: "2025-10-31T10:30:00Z"
                insufficientBalance:
                  summary: 포인트 부족
                  value:
                    success: false
                    error:
                      code: "PAYMENT001"
                      message: "잔액이 부족합니다"
                      details:
                        currentBalance: 10000
                        requiredAmount: 29000
                    timestamp: "2025-10-31T10:30:00Z"
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: 주문 상세 조회
      description: 특정 주문의 상세 정보를 조회합니다.
      operationId: getOrder
      parameters:
        - name: orderId
          in: path
          required: true
          description: 주문 ID
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{orderId}/cancel:
    post:
      tags:
        - Orders
      summary: 주문 취소
      description: 주문을 취소하고 재고/포인트/쿠폰을 복원합니다. (제조 시작 전에만 가능)
      operationId: cancelOrder
      parameters:
        - name: orderId
          in: path
          required: true
          description: 주문 ID
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderRequest'
            example:
              reason: "주문 실수"
      responses:
        '200':
          description: 취소 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelOrderResponse'
        '400':
          description: 취소 불가능한 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "ORDER003"
                  message: "이미 제조가 시작되어 취소할 수 없습니다"
                timestamp: "2025-10-31T10:30:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 토큰을 사용한 인증
        
        예시: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        timestamp:
          type: string
          format: date-time
          example: "2025-10-31T10:30:00Z"
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: 에러 코드
              example: "PRODUCT001"
            message:
              type: string
              description: 에러 메시지
              example: "존재하지 않는 상품입니다"
            details:
              type: object
              description: 추가 에러 정보
        timestamp:
          type: string
          format: date-time
          example: "2025-10-31T10:30:00Z"
    Pagination:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
      properties:
        page:
          type: integer
          description: 현재 페이지 번호 (0부터 시작)
          example: 0
        size:
          type: integer
          description: 페이지 크기
          example: 10
        totalElements:
          type: integer
          description: 전체 항목 수
          example: 100
        totalPages:
          type: integer
          description: 전체 페이지 수
          example: 10

    # ==================== Product ====================
...
    TeaInfo:
      type: object
      required:
        - dayNumber
        - dayOfWeek
        - teaName
        - mainIngredients
        - expectedEffects
        - brewingGuide
      properties:
        dayNumber:
          type: integer
          minimum: 1
          maximum: 7
          example: 1
        dayOfWeek:
          type: string
          example: "월요일"
        teaName:
          type: string
          example: "레몬그라스 진저 티"
        mainIngredients:
          type: string
          example: "레몬그라스, 생강, 레몬밤"
        expectedEffects:
          type: string
          example: "활력 증진, 기분 전환"
        brewingGuide:
          type: string
          example: "80도 물 200ml에 3분"
...
    DeliveryAddress:
      type: object
      required:
        - recipient
        - phone
        - zipCode
        - address
        - addressDetail
      properties:
        recipient:
          type: string
          description: 수령인
          example: "김민지"
        phone:
          type: string
          description: 전화번호
          pattern: '^\d{3}-\d{4}-\d{4}$'
          example: "010-1234-5678"
        zipCode:
          type: string
          description: 우편번호
          example: "06234"
        address:
          type: string
          description: 주소
          example: "서울시 강남구 테헤란로 123"
        addressDetail:
          type: string
          description: 상세 주소
          example: "456호"
    CreateOrderRequest:
      type: object
      required:
        - deliveryAddress
      properties:
        couponId:
          type: integer
          format: int64
          description: 사용할 쿠폰 ID (선택사항)
          nullable: true
          example: 1
        deliveryAddress:
          $ref: '#/components/schemas/DeliveryAddress'
    OrderItemInfo:
      type: object
      required:
        - product
        - combination
        - quantity
        - price
      properties:
        product:
          type: string
          example: "기본 7일 박스"
        combination:
          type: string
          example: "피로 + 활력 + 시트러스"
        quantity:
          type: integer
          example: 1
        price:
          type: integer
          example: 29000
    PaymentInfo:
      type: object
      required:
        - totalAmount
        - discountAmount
        - finalAmount
        - method
      properties:
        totalAmount:
          type: integer
          description: 총 상품 금액
          example: 29000
        discountAmount:
          type: integer
          description: 할인 금액
          example: 14500
        finalAmount:
          type: integer
          description: 최종 결제 금액
          example: 14500
        method:
          type: string
          enum: [BALANCE]
          example: "BALANCE"
    DeliveryInfo:
      type: object
      required:
        - recipient
        - phone
        - address
        - estimatedDeliveryDate
      properties:
        recipient:
          type: string
          example: "김민지"
        phone:
          type: string
          example: "010-1234-5678"
        address:
          type: string
          example: "서울시 강남구 테헤란로 123 456호"
        estimatedDeliveryDate:
          type: string
          format: date
          example: "2025-11-03"
        trackingNumber:
          type: string
          nullable: true
          example: null
    CreateOrderResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - orderId
                - orderNumber
                - items
                - payment
                - delivery
                - status
                - orderedAt
              properties:
                orderId:
                  type: integer
                  format: int64
                  example: 1
                orderNumber:
                  type: string
                  example: "ORD-20251031-001"
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderItemInfo'
                payment:
                  $ref: '#/components/schemas/PaymentInfo'
                delivery:
                  $ref: '#/components/schemas/DeliveryInfo'
                status:
                  type: string
                  enum: [PENDING, PAID, PREPARING, SHIPPED, DELIVERED, CANCELLED]
                  example: "PAID"
                orderedAt:
                  type: string
                  format: date-time
                  example: "2025-10-31T14:30:00Z"
    OrderSummary:
      type: object
      required:
        - orderId
        - orderNumber
        - totalItems
        - finalAmount
        - status
        - orderedAt
      properties:
        orderId:
          type: integer
          format: int64
          example: 1
        orderNumber:
          type: string
          example: "ORD-20251031-001"
        totalItems:
          type: integer
          example: 2
        finalAmount:
          type: integer
          example: 29000
        status:
          type: string
          enum: [PENDING, PAID, PREPARING, SHIPPED, DELIVERED, CANCELLED]
          example: "PAID"
        orderedAt:
          type: string
          format: date-time
          example: "2025-10-31T14:30:00Z"
    OrderListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - content
                - pagination
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderSummary'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    OrderItemDetail:
      type: object
      required:
        - product
        - combination
        - quantity
        - weeklyTeas
      properties:
        product:
          type: object
          properties:
            name:
              type: string
              example: "기본 7일 박스"
            price:
              type: integer
              example: 29000
        combination:
          type: object
          properties:
            condition:
              type: string
              example: "피로"
            mood:
              type: string
              example: "활력"
            scent:
              type: string
              example: "시트러스"
        quantity:
          type: integer
          example: 1
        weeklyTeas:
          type: array
          items:
            $ref: '#/components/schemas/TeaInfo'
          minItems: 7
          maxItems: 7
    OrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - orderId
                - orderNumber
                - status
                - items
                - payment
                - delivery
                - orderedAt
              properties:
                orderId:
                  type: integer
                  format: int64
                  example: 1
                orderNumber:
                  type: string
                  example: "ORD-20251031-001"
                status:
                  type: string
                  enum: [PENDING, PAID, PREPARING, SHIPPED, DELIVERED, CANCELLED]
                  example: "PAID"
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderItemDetail'
                payment:
                  allOf:
                    - $ref: '#/components/schemas/PaymentInfo'
                    - type: object
                      properties:
                        paidAt:
                          type: string
                          format: date-time
                          example: "2025-10-31T14:30:15Z"
                delivery:
                  $ref: '#/components/schemas/DeliveryInfo'
                orderedAt:
                  type: string
                  format: date-time
                  example: "2025-10-31T14:30:00Z"
    CancelOrderRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          description: 취소 사유
          minLength: 1
          maxLength: 500
          example: "주문 실수"
    CancelOrderResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - orderId
                - status
                - refund
                - message
              properties:
                orderId:
                  type: integer
                  format: int64
                  example: 1
                status:
                  type: string
                  enum: [CANCELLED]
                  example: "CANCELLED"
                refund:
                  type: object
                  properties:
                    amount:
                      type: integer
                      example: 14500
                    method:
                      type: string
                      example: "BALANCE"
                    refundedAt:
                      type: string
                      format: date-time
                      example: "2025-10-31T15:00:00Z"
                message:
                  type: string
                  example: "주문이 취소되었습니다"

    # ==================== Coupon ====================
...
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "유효하지 않은 요청입니다"
            timestamp: "2025-10-31T10:30:00Z"
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "인증이 필요합니다"
            timestamp: "2025-10-31T10:30:00Z"
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "FORBIDDEN"
              message: "접근 권한이 없습니다"
            timestamp: "2025-10-31T10:30:00Z"
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "요청한 리소스를 찾을 수 없습니다"
            timestamp: "2025-10-31T10:30:00Z"
...